Refact platform specific connections
