+++ package.json
@@ -588,15 +588,15 @@
   },
   "dependencies": {
     "@lit-labs/observers": "^2.0.2",
-    "@malloydata/db-bigquery": "0.0.169",
-    "@malloydata/db-duckdb": "0.0.169",
-    "@malloydata/db-postgres": "0.0.169",
-    "@malloydata/db-snowflake": "0.0.169",
-    "@malloydata/db-trino": "0.0.169",
-    "@malloydata/malloy": "0.0.169",
-    "@malloydata/malloy-sql": "0.0.169",
-    "@malloydata/render": "0.0.169",
-    "@malloydata/syntax-highlight": "0.0.169",
+    "@malloydata/db-bigquery": "0.0.176",
+    "@malloydata/db-duckdb": "0.0.176",
+    "@malloydata/db-postgres": "0.0.176",
+    "@malloydata/db-snowflake": "0.0.176",
+    "@malloydata/db-trino": "0.0.176",
+    "@malloydata/malloy": "0.0.176",
+    "@malloydata/malloy-sql": "0.0.176",
+    "@malloydata/render": "0.0.176",
+    "@malloydata/syntax-highlight": "0.0.176",
     "@vscode/webview-ui-toolkit": "^1.2.1",
     "duckdb": "1.0.0",
     "lit": "^3.0.2",
