+++ src/server/connections/browser/duckdb_wasm_connection.ts
